requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
faiss-cpu==1.7.3
Flask==2.2.3
frozenlist==1.3.3
html2text==2020.1.16
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.123
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.2
packaging==23.0
paper-qa==0.1.0
pycryptodome==3.17
pydantic==1.10.7
pypdf==3.7.0
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
SQLAlchemy==1.4.47
tenacity==8.2.2
tiktoken==0.3.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
